// src/timeline/timeline.enum.ts
export enum TimelineAction {
  VOTE_CAST = 'vote_cast',
  VOTE_REVOKED = 'vote_revoked',
  ELECTION_STARTED = 'election_started',
  ELECTION_ENDED = 'election_ended',
  ELECTION_CREATED = 'election_created',
  ELECTION_DELETED = 'election_deleted',
  ELECTION_UPDATED = 'election_updated',
  ELECTION_READINESS_VIEWED = 'election_readiness_viewed',
  ELECTION_WIZARD_UPDATED = 'election_wizard_updated',
  ELECTION_SECTION_APPROVED = 'election_section_approved',
  ELECTION_SECTION_REJECTED = 'election_section_rejected',
  ELECTION_AUDIT_PACKAGE_EXPORTED = 'election_audit_package_exported',
  ELECTION_DEMO_STARTED = 'election_demo_started',
  ELECTION_DEMO_CLEARED = 'election_demo_cleared',
  ELECTION_DEMO_COMPLETED = 'election_demo_completed',
  ELECTION_PRODUCTION_STARTED = 'election_production_started',
  ELECTION_CLOSED = 'election_closed',
  ELECTION_SCRUTINY_PUBLISHED = 'election_scrutiny_published',
  ELECTION_COMMUNICATION_TEMPLATES_UPDATED = 'election_communication_templates_updated',
  ELECTION_SUPERVISION_MEMBERS_UPDATED = 'election_supervision_members_updated',
  VOTER_REGISTERED = 'voter_registered',
  VOTER_UNREGISTERED = 'voter_unregistered',
  VOTER_UPDATED = 'voter_updated',
  VOTER_INVITATION_SENT = 'voter_invitation_sent',
  VOTER_INVITATION_FAILED = 'voter_invitation_failed',
  VOTER_REMINDER_SENT = 'voter_reminder_sent',
  VOTER_REMINDER_FAILED = 'voter_reminder_failed',
  OPTION_ADDED = 'option_added',
  OPTION_REMOVED = 'option_removed',
  OPTION_UPDATED = 'option_updated',
  VOTE_COUNTED = 'vote_counted',
  VOTE_UNCOUNTED = 'vote_uncounted',
  VOTER_LOGGED_IN = 'voter_logged_in',
  VOTER_LOGGED_OUT = 'voter_logged_out',
  VOTER_PROFILE_VIEWED = 'voter_profile_viewed',
  voter_profile_viewed = 'voter_profile_viewed',
  VOTER_PROFILE_UPDATED = 'voter_profile_updated',
  VOTER_PASSWORD_CHANGED = 'voter_password_changed',
  VOTER_EMAIL_VERIFIED = 'voter_email_verified',
  VOTER_EMAIL_UPDATED = 'voter_email_updated',
  VOTER_CODE_GENERATED = 'voter_code_generated',
  VOTER_CODE_UPDATED = 'voter_code_updated',
  VOTER_CODE_DELETED = 'voter_code_deleted',
  VOTER_CODE_USED = 'voter_code_used',
  VOTER_CODE_EXPIRED = 'voter_code_expired',
  VOTER_CODE_INVALIDATED = 'voter_code_invalidated',
  VOTER_CODE_REGENERATED = 'voter code_regenerated',
  VOTER_CODE_RESET = 'voter_code_reset',
  VOTER_CODE_VERIFIED = 'voter_code_verified',
  VOTER_CODE_UNVERIFIED = 'voter_code_unverified',
  VOTER_CODE_USED_ONCE = 'voter_code_used_once',
  VOTER_CODE_USED_MULTIPLE = 'voter_code_used_multiple',
  VOTER_CODE_USED_IN_SESSION = 'voter_code_used_in_session',
  VOTER_CODE_USED_OUTSIDE_SESSION = 'voter_code_used_outside_session',
  VOTER_CODE_USED_IN_ELECTION = 'voter_code_used_in_election',
  VOTER_CODE_USED_IN_VOTE = 'voter_code_used_in_vote',
  VOTER_CODE_USED_IN_OPTION = 'voter_code_used_in_option',

  // GDPR Audit Actions
  DATA_EXPORTED = 'data_exported',
  DATA_RECTIFIED = 'data_rectified',
  DATA_DELETED = 'data_deleted',
  CONSENT_GIVEN = 'consent_given',
  CONSENT_WITHDRAWN = 'consent_withdrawn',
  BREACH_DETECTED = 'breach_detected',
  BREACH_AUTHORITY_NOTIFIED = 'breach_authority_notified',
  BREACH_SUBJECTS_NOTIFIED = 'breach_subjects_notified',
  COOKIE_CONSENT_UPDATED = 'cookie_consent_updated',
  DPIA_CREATED = 'dpia_created',
  DPIA_APPROVED = 'dpia_approved',
  DPA_REGISTERED = 'dpa_registered',
  PROCESSING_RECORD_CREATED = 'processing_record_created',
  PROCESSING_RECORD_UPDATED = 'processing_record_updated',
}